requirements.txt
# Core scientific computing
numpy==2.4.3

# Data analysis
pandas==3.0.1

# Scientific computing
scipy==1.17.1

# Plotting and visualization
matplotlib==3.10.8

# Image processing
Pillow==12.1.1
opencv-python-headless==4.13.0.92
tifffile==2026.3.3

# PDF processing
reportlab==4.0.7
PyMuPDF==1.27.2
pypdf==6.8.0

# Utilities
python-dateutil==2.9.0.post0
